+++ src/compas/robots/__init__.py
@@ -9,9 +9,6 @@ robots
 Model
 =====
 
-.. inheritance-diagram:: RobotModel Joint Link ToolModel Configuration
-    :parts: 1
-
 The root of the model is the :class:`RobotModel` class, which
 describes a robot consisting of a set of link elements, and a set of joint
 elements connecting the links together.
@@ -30,9 +27,6 @@ elements connecting the links together.
 Geometric description
 =====================
 
-.. inheritance-diagram:: Geometry MeshDescriptor Material Texture Color
-    :parts: 1
-
 The robot itself as well as its links can be geometrically described
 using the following classes.
 
@@ -50,9 +44,6 @@ using the following classes.
 Link
 ====
 
-.. inheritance-diagram:: Visual Collision Inertial Mass Inertia
-    :parts: 1
-
 The link is described as a rigid body with inertial, visual and collision values.
 
 .. autosummary::
@@ -68,9 +59,6 @@ The link is described as a rigid body with inertial, visual and collision values
 
 Joint
 =====
-
-.. inheritance-diagram:: ParentLink ChildLink Calibration Dynamics Limit Axis Mimic SafetyController
-    :parts: 1
 
 The joint describes the kinematics and dynamics of the robot's joint.
 
@@ -90,9 +78,6 @@ The joint describes the kinematics and dynamics of the robot's joint.
 
 Resources
 =========
-
-.. inheritance-diagram:: AbstractMeshLoader DefaultMeshLoader GithubPackageMeshLoader LocalPackageMeshLoader
-    :parts: 1
 
 Model descriptions usually do not contain embedded geometry information but only
 descriptions, filenames or URLs for externally hosted resources.